requirements.txt
langchain
langchain-google-genai
qdrant-client
PyMuPDF
pdfplumber
pytesseract
pillow
opencv-python
numpy
streamlit
sentence-transformers
transformers
torch
accelerate
tomli